LayoutTests/imported/w3c/web-platform-tests/shadow-dom/input-element-list-expected.txt



FAIL Input element's list attribute should point to the datalist element. assert_equals: expected (object) null but got (undefined) undefined
FAIL Input element's list attribute should point to the datalist element in Shadow DOM. assert_equals: expected (object) null but got (undefined) undefined
